Enable SSE instructions during boot; Switched to qemu-x86_64 for default SSE support; Implemented xv6 physical page allocator

+++ src/traps.rs
@@ -25,7 +25,7 @@ pub const T_SIMDERR: u8 = 19; // SIMD floating point error
 // These are arbitrarily chosen, but with care not to overlap
 // processor defined exceptions or interrupt vectors.
 pub const T_SYSCALL: u8 = 64; // system call
-pub const T_DEFAULT: u8 = 500; // catchall
+pub const T_DEFAULT: u16 = 500; // catchall
 
 pub const T_IRQ0: u8 = 32; // IRQ 0 corresponds to int T_IRQ
